Fix: more codacy errors in educators
