+++ src/components/Evaluation.js
@@ -1,4 +1,4 @@
-import { useState } from 'react';
+import { useState,useEffect } from 'react';
 import axios from 'axios';
 import { toast} from 'react-toastify';
 
@@ -132,6 +132,57 @@ export default function EducatorEvaluationCommon() {
     // Return the grade of the most recent evaluation
     return filteredEvaluations[0].grade;
   };
+  useEffect(() => {
+    const id = localStorage.getItem('userId');
+    setUserId(id);
+  }, []);
+
+  useEffect(() => {
+    if (userId) {
+      axios.get(`${API_ENDPOINT}user/`)
+        .then(response => {
+          const user = response.data.find(user => user.id == userId);
+          if (user) {
+            setEducatorId(user.educator);
+          }
+        });
+    }
+  }, [userId]);
+
+  useEffect(() => {
+    if (educatorId) {
+      axios.get(`${API_ENDPOINT}lesson-event/`)
+        .then(response => {
+          const event = response.data.find(event => event.educators.includes(educatorId));
+          if (event) {
+            setKids(event.attendees);
+          }
+        });
+    }
+  }, [educatorId]);
+
+  useEffect(() => {
+    if (kids) {
+      axios.get(`${API_ENDPOINT}student/`)
+        .then(response => {
+          const matchingStudents = response.data.filter(student => kids.includes(student.id));
+          setStudents(matchingStudents);
+        });
+    }
+  }, [kids]);
+
+  useEffect(() => {
+    if (selectedStudent) {
+      axios.get(`${API_ENDPOINT}user/`)
+        .then(response => {
+          const user = response.data.find(user => user.family == selectedStudent.family);
+          if (user) {
+            setEmail(user.email);
+            setPhone(user.phone);
+          }
+        });
+    }
+  }, [selectedStudent]);
 
 
   return {
